requirements.txt
requests>=2.32.0
psycopg2-binary>=2.9.9
beautifulsoup4>=4.12.3

# ML & data
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.5.1
pyarrow>=16.1.0
joblib>=1.4.2

